refactor(FE): `useFakemonsByGym` hook
