+++ frontend/src/components/Battle/Battle.js
@@ -3,13 +3,10 @@ import Button from "react-bootstrap/Button";
 import Card from "react-bootstrap/Card";
 import { Link } from "react-router-dom";
 import { useParams } from "react-router-dom";
-import {
-  useFakemonsByGym,
-  useFakemonsByUser,
-  useIsRegistered,
-} from "../../hooks/queries";
+import { useFakemonsByUser, useIsRegistered } from "../../hooks/queries";
 import useAuth from "../../hooks/useAuth";
 import useCurrentBattle from "../../hooks/useCurrentBattle";
+import useFakemonsByGym from "../../hooks/useFakemonsByGym";
 import useGyms from "../../hooks/useGyms";
 import Alert from "../Alert";
 import ImportWalletAlert from "../ImportWalletAlert";
